upgrade react-native-svg move bars rendering into MessageDetailComponent
