+++ package.json
@@ -70,7 +70,7 @@
     "react-native-qrcode-scanner": "^1.1.0",
     "react-native-sha256": "1.2.3",
     "react-native-splash-screen": "^3.2.0",
-    "react-native-svg": "^6.5.3",
+    "react-native-svg": "^12.1.0",
     "react-native-text-input-mask": "^0.9.0",
     "react-native-touch-id": "^4.4.1",
     "react-native-vector-icons": "^6.6.0",
